KOA-5122 clean-up docs snapshot tests
